smallchat/requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
certifi==2017.4.17
channels==1.1.6
chardet==3.0.4
click==6.7
constantly==15.1.0
daphne==1.3.0
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
flockos==1.2.0
haikunator==2.1.0
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
msgpack-python==0.4.8
psycopg2==2.7.2
PyJWT==1.5.2
pytz==2017.2
redis==2.10.5
requests==2.18.1
rq==0.8.0
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
wheel==0.24.0
zope.interface==4.4.2
